requirements.txt
beautifulsoup4==4.12.3
boto3==1.34.84
botocore==1.34.84
bs4==0.0.1
certifi==2024.2.2
chardet==3.0.4
idna==2.10
jmespath==1.0.1
loguru==0.5.3
numpy==1.24.4
pandas==1.3.1
PyMySQL==1.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.24.0
s3transfer==0.10.1
six==1.16.0
soupsieve==2.5
tqdm==4.66.2
urllib3==1.25.11